+++ src/app/app.module.ts
@@ -12,6 +12,7 @@ import { mockApiServices } from 'app/mock-api';
 import { LayoutModule } from 'app/layout/layout.module';
 import { AppComponent } from 'app/app.component';
 import { appRoutes } from 'app/app.routing';
+import { PromptsModule } from './modules/prompts/prompts.modules';
 
 const routerConfig: ExtraOptions = {
     preloadingStrategy       : PreloadAllModules,
@@ -37,6 +38,7 @@ const routerConfig: ExtraOptions = {
 
         // Layout module of your application
         LayoutModule,
+        PromptsModule,
 
         // 3rd party modules that require global configuration via forRoot
         MarkdownModule.forRoot({})
