+++ src/app/modules/abm/abm-usuarios/abm-usuarios.component.scss
@@ -1,0 +1,6 @@
+.content-scroll {
+    flex-shrink: 1; 
+    flex-grow: 1; 
+    position: relative; 
+    overflow-y:scroll;
+}
