--- src/app/modules/abm/abm-usuarios/dialog/abm-usuarios-dialog.component.html
@@ -1,111 +1,0 @@
-<fuse-alert [type]="'success'" *ngIf="showSuccess">
-    <span fuseAlertTitle>Cambios realizados</span>
-</fuse-alert>
-<fuse-alert [type]="'error'" *ngIf="showError">
-    <span fuseAlertTitle>No se pudieron realizar los cambios</span>
-</fuse-alert>
-<mat-tab-group>
-    <mat-tab label="Usuario">
-        <div class="grid">
-            <div>
-                <h1 class="text-2xl font-semibold mb-5">Usuario</h1>
-            </div>
-            <mat-form-field class="fuse-mat-textarea">
-                <mat-label>Usuario</mat-label>
-                <input matInput [(ngModel)]="data.row.username">
-            </mat-form-field>
-            <mat-form-field class="fuse-mat-textarea">
-                <mat-label>Nombre</mat-label>
-                <input matInput [(ngModel)]="data.row.name">
-            </mat-form-field>
-            <mat-form-field class="fuse-mat-textarea">
-                <mat-label>Apellido</mat-label>
-                <input matInput [(ngModel)]="data.row.lastname">
-            </mat-form-field>
-            <section class="fuse-mat-textarea">
-                <mat-checkbox [(ngModel)]="data.row.enabled">
-                    Habilitado
-                </mat-checkbox>
-            </section>
-        </div>
-    </mat-tab>
-    <mat-tab label="Perfiles Asociados">
-        <div class="grid">
-            <div>
-                <h1 class="text-2xl font-semibold mb-5">Perfiles Asociados</h1>
-            </div>
-            <div class="grid grid-cols-5 gap-4">
-                <div class="col-span-2">
-                    <mat-form-field class="w-full">
-                        <mat-label>Perfiles Disponibles</mat-label>
-                        <input matInput (keyup)="onKeyDisponibles($event.target.value)" [(ngModel)]="inputDisponibles">
-                    </mat-form-field>
-                    <div style="height: 400px; overflow: auto; overflow-x: hidden;" *ngIf="perfilesDisponibles.length > 0">
-                        <table mat-table [dataSource]="searchPerfiles">
-                            <ng-container matColumnDef="select">
-                                <th mat-header-cell *matHeaderCellDef>
-                                  <mat-checkbox (change)="$event ? masterToggle(1) : null"
-                                                [checked]="selection.hasValue() && isAllSelected(1)"
-                                                [indeterminate]="selection.hasValue() && !isAllSelected(1)">
-                                  </mat-checkbox>
-                                </th>
-                                <td mat-cell *matCellDef="let row">
-                                  <mat-checkbox (click)="$event.stopPropagation()"
-                                                (change)="$event ? selection.toggle(row) : null"
-                                                [checked]="selection.isSelected(row)">
-                                  </mat-checkbox>
-                                </td>
-                            </ng-container>
-                            <ng-container matColumnDef="name">
-                                <th mat-header-cell *matHeaderCellDef> Nombre </th>
-                                <td mat-cell *matCellDef="let element">{{ element.name }}</td>
-                            </ng-container>
-                            <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-                            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-                        </table>
-                    </div>
-                </div>
-                <div class="col-span-1 text-center grid" style="align-items: center">
-                    <button mat-raised-button (click)="add()" [disabled]="selection.selected.length == 0" class="justify-self-center">Agregar</button>
-                    <button mat-raised-button (click)="remove()" [disabled]="selectionIncluidos.selected.length == 0" class="justify-self-center">Quitar</button>
-                </div>
-                <div class="col-span-2">
-                    <mat-form-field class="w-full">
-                        <mat-label>Perfiles Incluidos</mat-label>
-                        <input matInput (keyup)="onKeyIncluidos($event.target.value)" [(ngModel)]="inputIncluidos">
-                    </mat-form-field>
-                    <div style="height: 400px; overflow: auto; overflow-x: hidden;" *ngIf="perfilesIncluidos.length > 0">
-                        <table mat-table [dataSource]="searchPerfilesIncluidos">
-                            <ng-container matColumnDef="select">
-                                <th mat-header-cell *matHeaderCellDef>
-                                  <mat-checkbox (change)="$event ? masterToggle(2) : null"
-                                                [checked]="selectionIncluidos.hasValue() && isAllSelected(2)"
-                                                [indeterminate]="selectionIncluidos.hasValue() && !isAllSelected(2)">
-                                  </mat-checkbox>
-                                </th>
-                                <td mat-cell *matCellDef="let row">
-                                  <mat-checkbox (click)="$event.stopPropagation()"
-                                                (change)="$event ? selectionIncluidos.toggle(row) : null"
-                                                [checked]="selectionIncluidos.isSelected(row)">
-                                  </mat-checkbox>
-                                </td>
-                            </ng-container>
-                            <ng-container matColumnDef="name">
-                                <th mat-header-cell *matHeaderCellDef> Nombre </th>
-                                <td mat-cell *matCellDef="let element">{{ element.name }}</td>
-                            </ng-container>
-                            <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-                            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-                        </table>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </mat-tab>
-</mat-tab-group>
-<div class="grid">
-    <div class="mt-2 flex justify-around">
-        <button mat-raised-button color="warn" (click)="delete()">Eliminar</button>
-        <button mat-raised-button color="primary" (click)="edit()">Editar</button>
-    </div>
-</div>
