+++ src/app/modules/abm/abm-usuarios/grilla/abm-usuarios-grilla.component.html
@@ -1,0 +1,46 @@
+<div class="flex flex-col flex-auto min-w-0">
+    <!-- Main -->
+    <div class="flex-auto p-6 sm:p-10">
+        <!-- CONTENT GOES HERE -->
+        <div class="h-400 min-h-400 max-h-400">
+            <table mat-table [dataSource]="usuarios" style="width: 100%;" class="border-2 rounded-2xl">
+                <ng-container matColumnDef="usuario">
+                    <th mat-header-cell *matHeaderCellDef> Usuario </th>
+                    <td mat-cell *matCellDef="let element"> {{element.username}} </td>
+                </ng-container>
+                <ng-container matColumnDef="nombre">
+                    <th mat-header-cell *matHeaderCellDef> Nombre </th>
+                    <td mat-cell *matCellDef="let element"> {{element.name}} </td>
+                </ng-container>
+                <ng-container matColumnDef="apellido">
+                    <th mat-header-cell *matHeaderCellDef> Apellido </th>
+                    <td mat-cell *matCellDef="let element"> {{element.lastname}} </td>
+                </ng-container>
+                <ng-container matColumnDef="email">
+                    <th mat-header-cell *matHeaderCellDef> Email </th>
+                    <td mat-cell *matCellDef="let element"> {{element.email}} </td>
+                </ng-container>
+                <ng-container matColumnDef="perfil">
+                    <th mat-header-cell *matHeaderCellDef> Perfil </th>
+                    <td mat-cell *matCellDef="let element"> {{element.profiles[0].name}} </td>
+                </ng-container>
+                <ng-container matColumnDef="acciones">
+                    <th mat-header-cell *matHeaderCellDef></th>
+                    <td mat-cell *matCellDef="let row">
+                        <button>
+                            <mat-icon svgIcon="heroicons_solid:pencil-alt" [routerLink]="['../user', row.id]" (click)="openUser(1, row)"></mat-icon>
+                        </button>
+                        <button>
+                            <mat-icon svgIcon="heroicons_solid:eye" [routerLink]="['../user', row.id]" (click)="openUser(2, row)"></mat-icon>
+                        </button>
+                        <button>
+                            <mat-icon svgIcon="heroicons_solid:document-remove" (click)="delete(row)"></mat-icon>
+                        </button>
+                    </td>
+                </ng-container>
+                <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+                <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+            </table>
+        </div>
+    </div>
+</div>
