+++ src/app/modules/abm/abm-usuarios/user/abm-usuarios-user.component.scss
@@ -1,0 +1,14 @@
+.lista-permisos {
+    max-height: fit-content;
+    overflow: auto; 
+    overflow-x: hidden;
+}
+
+.buttons-grid {
+    flex-direction: row;
+}
+
+.buttons-grid button {
+    flex-shrink: 0;
+    margin-left: 5px;
+}
