+++ src/app/modules/auth/sign-in/sign-in.component.ts
@@ -82,7 +82,7 @@ export class AuthSignInComponent implements OnInit
         }
 
         this.loginService.login(model).subscribe(res => {
-            const redirectURL = this._activatedRoute.snapshot.queryParamMap.get('usuarios') || '/usuarios';
+            const redirectURL = this._activatedRoute.snapshot.queryParamMap.get('usuarios') || '/usuarios/grid';
             this._router.navigateByUrl(redirectURL);
         },
         err => {
