+++ src/app/shared/services/user.service.ts
@@ -13,6 +13,8 @@ import { Injectable } from "@angular/core";
 
 export class UserService {
 
+    private mode: string;
+
     constructor(
         private http: HttpClient,
         private handler: HttpBackend) {
@@ -41,5 +43,13 @@ export class UserService {
 
     public updateUserValidation(user: User): Observable<Respuesta> {
         return this.http.post<Respuesta>(`${environment.server}users/update/validation/`, user)
+    }
+
+    public getMode() {
+        return this.mode;
+    }
+
+    public setMode(mode: string) {
+        this.mode = mode;
     }
 }
